chore: move init into its own module
