+++ examples/ipns-resolve.rs
@@ -9,9 +9,9 @@ async fn main() -> anyhow::Result<()> {
     use std::str::FromStr;
 
     use clap::Parser;
+    use rust_ipfs::builder::UninitializedIpfsDefault as UninitializedIpfs;
     use rust_ipfs::Ipfs;
     use rust_ipfs::IpfsPath;
-    use rust_ipfs::UninitializedIpfsDefault as UninitializedIpfs;
 
     tracing_subscriber::fmt::init();
