+++ examples/local-node.rs
@@ -1,7 +1,7 @@
 use rust_ipfs::Ipfs;
 
+use rust_ipfs::builder::UninitializedIpfs;
 use rust_ipfs::Keypair;
-use rust_ipfs::UninitializedIpfs;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
