+++ examples/request-response.rs
@@ -2,7 +2,7 @@ use bytes::Bytes;
 use connexa::behaviour::request_response::RequestResponseConfig;
 use futures::StreamExt;
 
-use rust_ipfs::UninitializedIpfsDefault as UninitializedIpfs;
+use rust_ipfs::builder::UninitializedIpfsDefault as UninitializedIpfs;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
