+++ examples/swarm_events.rs
@@ -1,6 +1,6 @@
 use connexa::prelude::swarm::SwarmEvent;
+use rust_ipfs::builder::UninitializedIpfsDefault as UninitializedIpfs;
 use rust_ipfs::Ipfs;
-use rust_ipfs::UninitializedIpfsDefault as UninitializedIpfs;
 use std::time::Duration;
 
 #[tokio::main]
