+++ examples/unixfs-cat-readme.rs
@@ -1,6 +1,6 @@
 use std::str::FromStr;
 
-use rust_ipfs::UninitializedIpfsDefault as UninitializedIpfs;
+use rust_ipfs::builder::UninitializedIpfsDefault as UninitializedIpfs;
 use rust_ipfs::{Ipfs, IpfsPath};
 use tokio::io::AsyncWriteExt;
